+++ lib/src/class_internal/table.dart
